requirements.txt
numpy==1.26.4
pandas==2.2.1
Pillow==10.3.0
Requests==2.31.0
scikit_learn==1.4.1.post1
selenium==4.19.0
streamlit==1.32.0
tensorflow==2.16.1
tensorflow_intel==2.16.1
tqdm==4.66.2
